Resolving Draw Backs [1.0]
Resolving Few Draw Backs in Bus Info and Available Trip Option

+++ BusTicketManagementSystem/User_Controls/Bus_Info.cs
@@ -35,8 +35,6 @@ namespace BusTicketManagementSystem.User_Controls
             string busClass = busClassComboBox.Text;
             string busStatus = busStatusComboBox.Text;
 
-            bool busExist = false;
-
             //Checking if any field is empty...
             if (string.IsNullOrEmpty(busNumber) || string.IsNullOrEmpty(busClass) || string.IsNullOrEmpty(busStatus))
             {
@@ -44,67 +42,95 @@ namespace BusTicketManagementSystem.User_Controls
             }
             else
             {
-                //Checking Bus number that if it's already exist or not
-                var reader = db.GetData("SELECT bus_number FROM Bus");
-                while (reader.Read())
+                //Checking if the bus already added in the List
+                if(busExistance("SELECT bus_number FROM Bus", busNumber))
                 {
-                    if(busNumber == reader["bus_number"].ToString())
-                    {
-                        busExist = true;//this will help to know bus existance
-                        MessageBox.Show("Bus Already Exist", "WARNING", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                        break;
-                    }
+                    MessageBox.Show("Bus Already Exist", "WARNING", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 }
-
-                //IF bus number is unique then the below if block will INSERT the bus...
-                if(busExist == false)
+                else
                 {
-                    db.SetData("INSERT INTO Bus VALUES('"+busNumber+"', '"+busClass+"', '"+busStatus+"')");
-                    busNumberTextBox.Text = "";             
+                    db.SetData("INSERT INTO Bus VALUES('" + busNumber + "', '" + busClass + "', '" + busStatus + "')");
+                    busNumberTextBox.Text = "";
                     busStatusComboBox.SelectedIndex = -1;
                     busClassComboBox.SelectedIndex = -1;
                     MessageBox.Show("Bus Added Successfully", "SUCCESS", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
-                
+
             }
         }
 
-
+        //Fetching All data from Bus Table
         public void showAllBus()
         {
             var reader = db.GetData("SELECT * FROM Bus");
             populateGrid(reader);
+            //Checking if the scrollBar Exist
+            foreach (var scroll in busInfoGrid.Controls.OfType<VScrollBar>())
+            {
+                if (scroll.Visible)
+                {
+                    panel1.Visible = true;
+                }
+                else
+                {
+                    panel1.Visible = false;
+                }
+            }
         }
-  
+        
+
+        //Bus Info Edit & Delete Functionality
         private void busInfoGrid_CellClick(object sender, DataGridViewCellEventArgs e)
         {
+            //Delete...
             if(busInfoGrid.Columns[e.ColumnIndex].Name == "deleteColumn")
             {
                 string selectedBusNumber = busInfoGrid.Rows[e.RowIndex].Cells[0].Value.ToString();
                 DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete?", "DATA DELETION WARNING", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                
                 if (dialogResult == DialogResult.Yes)
                 {
-                    db.SetData("DELETE FROM Bus WHERE bus_number = '"+selectedBusNumber+"'");
-                    MessageBox.Show("Data Deleted Successfully", "SUCCESS", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    //If bus assigned for a trip delete operation not possible
+                    if(busExistance("SELECT bus_number FROM Trip", selectedBusNumber))
+                    {
+                        MessageBox.Show("Bus already assigned for a trip", "Can't Delete", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        db.SetData("DELETE FROM Bus WHERE bus_number = '" + selectedBusNumber + "'");
+                        MessageBox.Show("Data Deleted Successfully", "SUCCESS", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    }
                 }
             }
 
+            //Edit...
             if(busInfoGrid.Columns[e.ColumnIndex].Name == "editColumn")
             {
                 string selectedBusNumber = busInfoGrid.Rows[e.RowIndex].Cells[0].Value.ToString();
                 string selectedBusClass = busInfoGrid.Rows[e.RowIndex].Cells[1].Value.ToString();
                 string selectedBusStatus = busInfoGrid.Rows[e.RowIndex].Cells[2].Value.ToString();
-                Popups.EditBusInfo_Popup editBusInfo_Popup = new Popups.EditBusInfo_Popup(selectedBusNumber, selectedBusClass, selectedBusStatus);
-                editBusInfo_Popup.ShowDialog();
+
+                //If bus assigned for a trip UPDATE operation not possible
+                if (busExistance("SELECT bus_number FROM Trip", selectedBusNumber))
+                {
+                    MessageBox.Show("Bus already assigned for a trip", "Can't Update", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+                else
+                {
+                    Popups.EditBusInfo_Popup editBusInfo_Popup = new Popups.EditBusInfo_Popup(selectedBusNumber, selectedBusClass, selectedBusStatus);
+                    editBusInfo_Popup.ShowDialog();
+                }
             }
         }
 
+        //Search Box...
         private void searchBoxEvenet(object sender, EventArgs e)
         {
             var reader = db.GetData("SELECT * FROM Bus WHERE bus_number LIKE '"+searchBox.Text+"%'");
             populateGrid(reader);
         }
 
+        //AFunction Fetching data from database and populating Grid
         private void populateGrid(SqlDataReader reader)
         {
             busInfoGrid.Rows.Clear();
@@ -113,5 +139,23 @@ namespace BusTicketManagementSystem.User_Controls
                 busInfoGrid.Rows.Add(reader["bus_number"], reader["bus_class"], reader["bus_status"]);
             }
         }
+
+        //A function for checking if a bus is assigned for a trip or not
+        private bool busExistance(string query, string busNumber)
+        {
+            var reader = db.GetData(query);
+            bool existance = false;
+            while (reader.Read())
+            {
+                if(busNumber == reader["bus_number"].ToString())
+                {
+                    existance = true;
+                    break;
+                }
+            }
+            return existance;
+        }
+
+
     }
 }
